[TASK] remove function
