+++ steam/ext/csgo/protobufs/cstrike.py
@@ -117,15 +117,6 @@ class OperationalStatisticsPacket(betterproto.Message):
     values: "list[OperationalStatisticElement]" = betterproto.message_field(3)
 
 
-@dataclass(eq=False, repr=False)
-class PlayerRankingInfo(betterproto.Message):
-    account_id: int = betterproto.uint32_field(1)
-    rank_id: int = betterproto.uint32_field(2)
-    wins: int = betterproto.uint32_field(3)
-    rank_change: float = betterproto.float_field(4)
-    rank_type_id: int = betterproto.uint32_field(6)
-    tv_control: int = betterproto.uint32_field(7)
-
 
 @dataclass(eq=False, repr=False)
 class PlayerCommendationInfo(betterproto.Message):
@@ -217,6 +208,29 @@ class ScoreLeaderboardData(betterproto.Message):
     accountentries: "list[ScoreLeaderboardDataAccountEntries]" = betterproto.message_field(3)
     matchentries: "list[ScoreLeaderboardDataEntry]" = betterproto.message_field(5)
 
+
+
+@dataclass(eq=False, repr=False)
+class PlayerRankingInfo(GCProtobufMessage, msg=EMsg.ClientGetPlayerRankingInfoResponse):
+    account_id: int = betterproto.uint32_field(1)
+    rank_id: int = betterproto.uint32_field(2)
+    wins: int = betterproto.uint32_field(3)
+    rank_change: float = betterproto.float_field(4)
+    rank_type_id: int = betterproto.uint32_field(6)
+    tv_control: int = betterproto.uint32_field(7)
+    rank_window_stats: int = betterproto.uint64_field(8)
+    leaderboard_name: str = betterproto.string_field(9)
+    rank_if_win: int = betterproto.uint32_field(10)
+    rank_if_lose: int = betterproto.uint32_field(11)
+    rank_if_tie: int = betterproto.uint32_field(12)
+    per_map_rank: "list[PerMapRank]" = betterproto.message_field(13)
+    leaderboard_name_status: int = betterproto.uint32_field(14)
+
+@dataclass(eq=False, repr=False)
+class PerMapRank(betterproto.Message):
+    map_id: int = betterproto.uint32_field(1)
+    rank_id: int = betterproto.uint32_field(2)
+    wins: int = betterproto.uint32_field(3)
 
 @dataclass(eq=False, repr=False)
 class ScoreLeaderboardDataEntry(betterproto.Message):
@@ -718,6 +732,13 @@ class ClientRequestPlayersProfile(GCProtobufMessage, msg=EMsg.ClientRequestPlaye
     request_level: int = betterproto.uint32_field(4)
 
 
+#class ClientRequestPlayersProfilecs2(GCProtobufMessage, msg=EMsg.ClientRequestPlayersProfilecs2):
+#    request_id_deprecated: int = betterproto.uint32_field(1)
+#    account_ids_deprecated: list[int] = betterproto.uint32_field(2)
+#    account_id: int = betterproto.uint32_field(3)
+#    request_level: int = betterproto.uint32_field(4)
+
+
 @dataclass(eq=False, repr=False)
 class PlayersProfileProfile(betterproto.Message):
     account_id: int = betterproto.uint32_field(1)
@@ -744,6 +765,11 @@ class PlayersProfileProfile(betterproto.Message):
 class PlayersProfile(GCProtobufMessage, msg=EMsg.PlayersProfile):
     request_id: int = betterproto.uint32_field(1)
     account_profiles: "list[PlayersProfileProfile]" = betterproto.message_field(2)
+
+class PlayersProfilecs2(GCProtobufMessage, msg=EMsg.PlayersProfilecs2):
+    request_id: int = betterproto.uint32_field(1)
+    account_profiles_cs2: "list[PlayersProfileProfile]" = betterproto.message_field(2)
+
 
 
 @dataclass(eq=False, repr=False)
